# documentation/api/openapi.json
{
  "components": {
    "schemas": {
...
      "Error": {
        "description": "An error as a message, for access methods that carry errors in a body\n(REST, step 12): the same code and message as the gRPC status.",
        "properties": {
          "code": {
            "description": "One of the codes above.",
            "type": "string"
          },
          "message": {
            "description": "For people.",
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "description": "A session token from `POST /v1/auth/login`, or an API token.",
        "scheme": "bearer",
        "type": "http"
      },
      "sessionCookie": {
        "description": "The console's session (login with `cookie: true`). Requests other than GET and HEAD must also send the `X-Iwdb-Csrf` header.",
        "in": "cookie",
        "name": "iwdb_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "The `Database` trait over HTTP/JSON, with the messages of the gRPC contract (`proto/ironweaver_db/v1`) in their proto3 JSON form. See documentation/api/rest.md.",
    "license": {
      "identifier": "AGPL-3.0-only",
      "name": "AGPL-3.0-only (or a commercial license)"
    },
    "title": "Ironweaver DB REST API",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/openapi.json": {
      "get": {
        "operationId": "openapi",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "The OpenAPI document."
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "A request the server refuses: the HTTP status of its code (documentation/api/errors.md)."
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "A failure of the server: the HTTP status of its code (documentation/api/errors.md)."
          }
        },
        "security": [],
        "summary": "This API's OpenAPI document"
      }
    },
...
  },
  "security": [
    {
      "bearer": []
    },
    {
      "sessionCookie": []
    }
  ],
  "servers": [
    {
      "url": "http://127.0.0.1:7600"
    }
  ]
}